fix: low-severity bugs L-01 through L-04 and dead code cleanup D-07 through D-15
L-01: ChdFile.Dispose - dispose ThreadLocal values with trackAllValues:true
L-02: HashUtil.IsAllZero - add null check
L-03: ShowAbout - guard Application.Current?.MainWindow
L-04: HunkDebug - cast to int before array indexing

D-07: ChdHeaderInfo.IsAllZero - reuse Util.IsAllZeroArray
D-09: MapEntry compression constants - kept (used by MapCompressor)
D-11: MetadataWriter tag constants - kept (used by tests)
D-13: Extract duplicate NonSeekableStream to shared TestHelpers.cs
D-14: Extract duplicate RunChdman/ResolveChdmanPath to ChdmanHelper.cs (16 files)
D-15: Remove duplicate CompTypeConv/IsValidCodec tests from SecondCompressedTests.cs

+++ CHDSharpTestGen/Program.cs
@@ -137,7 +137,7 @@ internal static class Program
                     failures++;
                     var firstDiff = -1;
                     for (var i = 0; i < buffer.Length; i++)
-                        if (buffer[i] != raw[h * hdr.Blocksize + i])
+                        if (buffer[i] != raw[(int)(h * hdr.Blocksize) + i])
                         {
                             firstDiff = i;
                             break;
